+++ src/types/medias.d.ts
@@ -1,4 +1,4 @@
-export interface Jingle {
+export interface Media {
 	file: string,
 	gain: number
 }